tests: migrate TestReinitializeValues from System->Composition
